Additional backend refactoring
