+++ environment.yml
@@ -4,7 +4,7 @@ channels:
   - conda-forge
 
 dependencies:
-  - python=3.10
+  - python=3.8
   - celery=5.2.7
   - django=4.1
   - django-ninja=0.19.1
@@ -13,17 +13,17 @@ dependencies:
   - pyyaml=6.0
   - inflection=0.5.1
   - sqlalchemy=1.4.45
-  - djangorestframework=3.14
   - django-cors-headers==3.14
 
   - pip:
       - Authlib==1.2.1
-      - django-hydrothings==0.1.5
+      - hydroserver-sensorthings==0.1.8
       - django-ninja-jwt==5.2.5
       - hydroloader==0.1.12
       - sqlalchemy-json==0.5.0
       - geojson-pydantic==0.5.0
-      - djangorestframework-simplejwt==4.4.0
       - boto3==1.28.8
       - django-ses==3.5.0
       - django-storages==1.13.2
+      - django_ses==2.5.0
+      - djangorestframework-simplejwt==0.0.2
