+++ hydroserver/urls.py
@@ -3,13 +3,15 @@ from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import path, include
 from sites.api import api
+from hydroserver.views import index
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/', api.urls),
     path('api/data/', include('core.urls')),
     path('api/account/', include('accounts.urls')),
-    path('api/sensorthings/', include('sensorthings.urls')),
+    path('api/sensorthings/', include('stapi.urls')),
+    path('', index)
 ]
 
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
