+++ requirements.txt
@@ -1,7 +1,7 @@
 django==4.1
 django-ninja==0.20.0
 django-ninja-jwt==5.2.5
-django-hydrothings==0.1.7
+hydroserver-sensorthings==0.1.8
 hydroloader==0.1.12
 dj-database-url==1.3.0
 python-decouple==3.7
@@ -10,12 +10,11 @@ pyyaml==6.0
 psycopg2-binary==2.9.5
 pandas==1.5.3
 numpy==1.24.2
-djangorestframework==3.14
 django-cors-headers==3.14
-djangorestframework-simplejwt==5.2.2
 urllib3==1.26.6
 boto3==1.28.8
 botocore==1.31.15
 django-ses==2.5.0
 Authlib==1.2.1
 django-storages==1.13.2
+djangorestframework-simplejwt==0.0.2
